--- app/src/main/java/com/jyp/journeypiki/OnboardingEnum.kt
@@ -1,18 +1,0 @@
-package com.jyp.journeypiki
-
-import com.jyp.journeypiki.R
-
-
-enum class OnboardingEnum(
-    val titleRes: Int,
-    val imageRes: Int
-) {
-    ONBOARDING_01(
-        titleRes = R.string.onboarding_title01,
-        imageRes = R.drawable.image_onboarding01
-    ),
-    ONBOARDING_02(
-        titleRes = R.string.onboarding_title02,
-        imageRes = R.drawable.image_onboarding02
-    )
-}
