--- feature_question/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.jyp.feature_question">
-
-    <application>
-        <activity
-            android:name=".presentation.QuestionActivity"
-            android:exported="false" />
-    </application>
-
-</manifest>
