+++ settings.gradle
@@ -15,12 +15,12 @@ dependencyResolutionManagement {
 }
 rootProject.name = "JYPAndroid"
 include ':app'
-include ':core_network'
 include ':core_ui'
+include ':core_network'
+include ':feature_sign_in'
 include ':feature_main'
 include ':feature_my_page'
 include ':feature_my_journey'
 include ':feature_another_journey'
 include ':feature_planner'
-include ':feature_question'
 include ':feature_add_place'
